# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  description: |
    # Notification Service API Documentation
    
    A comprehensive notification service that supports multiple delivery channels (Email, SSE, Push) 
    with template management and user preferences.
    
    ## Features
    - 📧 **Email Notifications** - Send templated emails via SMTP
    - 🔔 **Real-time SSE** - Server-Sent Events for instant notifications
    - 📱 **Push Notifications** - (Coming soon)
    - 🎨 **Template Management** - Dynamic templates with variable substitution
    - ⚙️ **User Preferences** - Customizable notification settings
    - 🔄 **Event-Driven** - Redis Streams for reliable event processing
    - 🔁 **Retry Mechanism** - Automatic retry for failed notifications
    
    ## Authentication
    All endpoints (except health check) require JWT Bearer token authentication from Casdoor OAuth2.
    
    ## Base URL
    - **Development**: `http://localhost:8082/notification-service`
    - **Production**: `https://api.yourcompany.com/notification-service`

  version: 2.0.0
  contact:
    name: Notification Service Team
    email: support@examplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082/notification-service
    description: Development server
  - url: https://api.yourcompany.com/notification-service
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/redis/test/quick/user-registered:
    post:
      tags:
        - Events
      summary: Quick user registration test
      description: Quickly publish a user registration event (Admin only)
      operationId: quickUserRegistered
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
          example: 123
        - name: username
          in: query
          required: true
          schema:
            type: string
          example: johndoe
        - name: email
          in: query
          required: true
          schema:
            type: string
          example: john.doe@example.com
      responses:
        '200':
          description: Event published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
  /api/v1/redis/test/events/user-registered:
    post:
      tags:
        - Events
      summary: Publish user registered event
      description: Publish a detailed user registration event (Admin only)
      operationId: publishUserRegistered
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisteredEvent'
      responses:
        '200':
          description: Event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
  /api/v1/redis/test/events/session-completed:
    post:
      tags:
        - Events
      summary: Publish session completed event
      description: Publish a session completion event (Admin only)
      operationId: publishSessionCompleted
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCompletedEvent'
            example:
              userId: 123
              username: johndoe
              email: john.doe@example.com
              sessionId: SESSION-2025-001
              assessmentName: Spring Boot Advanced
              completionTime: "45 minutes"
              score: 95.5
              status: PASSED
      responses:
        '200':
          description: Event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
  /api/v1/redis/test/events/proctoring-violation:
    post:
      tags:
        - Events
      summary: Publish proctoring violation event
      description: Publish a proctoring violation event (Admin only)
      operationId: publishProctoringViolation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProctoringViolationEvent'
            example:
              userId: 123
              username: johndoe
              sessionId: SESSION-2025-001
              violationType: MULTIPLE_FACES_DETECTED
              severity: HIGH
              proctorIds: [1, 2, 3]
      responses:
        '200':
          description: Event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
  /api/v1/redis/test/events/assessment-published:
    post:
      tags:
        - Events
      summary: Publish assessment published event
      description: Publish an assessment publication event (Admin only)
      operationId: publishAssessmentPublished
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentPublishedEvent'
            example:
              assessmentId: ASSESS-2025-001
              assessmentName: Spring Boot Advanced Certification
              duration: 120
              dueDate: "2025-02-01T00:00:00Z"
              assignedUsers:
                - userId: 123
                  username: johndoe
                  email: john.doe@example.com
      responses:
        '200':
          description: Event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishResponse'
  /api/v1/redis/test/bulk/user-registered:
    post:
      tags:
        - Events
      summary: Bulk user registration test
      description: Publish multiple user registration events for testing (Admin only)
      operationId: bulkUserRegistered
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Bulk events published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  messageIds:
                    type: array
                    items:
                      type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from Casdoor OAuth2.
        
        To get a token, authenticate with Casdoor:
        ```
        POST https://casdoor.sapx.me/api/login/oauth/access_token
        Content-Type: application/x-www-form-urlencoded
        
        grant_type=password&
        username=YOUR_USERNAME&
        password=YOUR_PASSWORD&
        client_id=YOUR_CLIENT_ID&
        client_secret=YOUR_CLIENT_SECRET
        ```
  schemas:
...
    UserRegisteredEvent:
      type: object
      required:
        - userId
        - username
        - email
      properties:
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    SessionCompletedEvent:
      type: object
      required:
        - userId
        - username
        - sessionId
        - assessmentName
      properties:
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        sessionId:
          type: string
        assessmentName:
          type: string
        completionTime:
          type: string
        score:
          type: number
          format: double
        status:
          type: string
          enum: [PASSED, FAILED]
    ProctoringViolationEvent:
      type: object
      required:
        - userId
        - username
        - sessionId
        - violationType
      properties:
        userId:
          type: integer
        username:
          type: string
        sessionId:
          type: string
        violationType:
          type: string
          example: MULTIPLE_FACES_DETECTED
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        proctorIds:
          type: array
          items:
            type: integer
    AssessmentPublishedEvent:
      type: object
      required:
        - assessmentId
        - assessmentName
        - assignedUsers
      properties:
        assessmentId:
          type: string
        assessmentName:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        dueDate:
          type: string
          format: date-time
        assignedUsers:
          type: array
          items:
            type: object
            properties:
              userId:
                type: integer
              username:
                type: string
              email:
                type: string
    EventPublishResponse:
      type: object
      properties:
        success:
          type: boolean
        messageId:
          type: string
        stream:
          type: string
        event:
          type: object
        checkMailHog:
          type: string

    # ==================== PAGINATION ====================
...